+++ pubspec.yaml
@@ -40,7 +40,8 @@ dependencies:
   auto_size_text: ^3.0.0
   nfc_manager: ^3.3.0
   intl: ^0.18.1
-
+  provider: ^6.1.1
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -69,7 +70,7 @@ flutter:
     - assets/Escout Logo.png
     - assets/Email.png
     - assets/Hide Password.png
-    - assests/Images
+    # - assets/images
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
